Moves SteamPlayer to separate file due to circular import
